fix(arcjet): move Arcjet to middleware to prevent re-renders in layout
